refactor: move utils to helpers directory
